design: 카테고리 추가,수정 모달창 디자인변경 / 추가버튼 위치수정
