+++ src/components/common/modal/Modal.jsx
@@ -49,8 +49,7 @@ const ModalWrap = styled.div`
 `;
 
 const Contents = styled.div`
-  margin: 50px 30px;
-
+  margin: 40px 30px;
   img {
     margin-top: 60px;
     width: 300px;
